+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.6.0] - 2024-08-26
+
+- changed to using progress bar instead of simple spinner
+- added gum like spinner chars
+
 ## [0.5.0] - 2024-08-26
 
 - added --interactive flag
@@ -9,8 +14,6 @@
 ## [0.4.1] - 2024-08-25
 
 - making use of modules now
-
-WIP!
 
 ## [0.4.0] - 2024-07-28